Extract uploaded file normalization
